+++ Desktop Application/Basic/Python/Story-Generator/README.md
